sound and fixes
