+++ Content.Shared/Stories/Nightvision/SharedNightvisionSystem.cs
@@ -1,27 +1,22 @@
-using Content.Shared.Actions;
-
+using Robust.Shared.Audio.Systems;
+using Robust.Shared.Timing;
 namespace Content.Shared.Stories.Nightvision;
 
 public sealed class SharedNightvisionSystem : EntitySystem
 {
-    [Dependency] private readonly SharedActionsSystem _actions = default!;
+    [Dependency] private readonly IGameTiming _timing = default!;
+    [Dependency] private readonly SharedAudioSystem _audio = default!;
     public override void Initialize()
     {
         base.Initialize();
-        SubscribeLocalEvent<NightvisionComponent, ComponentStartup>(OnStartUp);
         SubscribeLocalEvent<NightvisionComponent, ToggleNightvisionEvent>(OnToggle);
-        SubscribeLocalEvent<NightvisionComponent, ComponentShutdown>(OnShutdown);
-    }
-    private void OnStartUp(EntityUid uid, NightvisionComponent component, ComponentStartup args)
-    {
-        _actions.AddAction(uid, ref component.ToggleActionEntity, component.ToggleAction);
     }
     private void OnToggle(EntityUid uid, NightvisionComponent component, ToggleNightvisionEvent args)
     {
+        if (!_timing.IsFirstTimePredicted)
+            return;
         component.Enabled = !component.Enabled;
-    }
-    private void OnShutdown(EntityUid uid, NightvisionComponent component, ComponentShutdown args)
-    {
-        Del(component.ToggleActionEntity);
+        if (component.Enabled)
+            _audio.PlayPredicted(component.ToggleOnSound, uid, uid);
     }
 }
