+++ Resources/Prototypes/Stories/Actions/nightvision.yml
@@ -2,11 +2,8 @@
   id: ToggleNightvisionAction
   name: Ночное зрение
   description: Переключить своё ночное зрение.
-  noSpawn: true
   components:
   - type: InstantAction
     useDelay: 5
     icon: { sprite: Stories/Clothing/Eyes/Glasses/nvg.rsi, state: icon }
-    itemIconStyle: NoItem
-    priority: -11
-    event: !type:ToggleNightvisionEvent {}
+    event: !type:ToggleNightvisionEvent
